refactor: relocate config GUI
